requirements.txt
Django==4.2.20
gunicorn==20.1.0
mysqlclient==2.2.7
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
sqlparse==0.5.3
rich==13.9.4
typing_extensions==4.12.2
PyYAML==6.0.2
arrow==1.3.0
asgiref==3.8.1
django-browser-reload==1.18.0
django-debug-toolbar==5.0.1
django-tailwind==3.6.0
python-slugify==8.0.4
six==1.17.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
urllib3==2.3.0

# dotenv==0.9.9
# idna==3.10
# Jinja2==3.1.6
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# Pygments==2.19.1
# PyMySQL==1.1.1
# binaryornot==0.4.4
# certifi==2025.1.31
# chardet==5.2.0
# charset-normalizer==3.4.1
# click==8.1.8
# cookiecutter==2.6.0